requirements.txt
asgiref==3.7.2
Django==4.2.5
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
phonenumbers==8.13.35
Pillow==10.1.0
psycopg2-binary==2.9.9
sqlparse==0.4.4
